pyproject.toml
[project]
name = "research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=2.1.10",
    "chardet>=5.2.0",
    "chonkie[openai,st]>=1.4.0",
    "openai>=2.6.0",
    "pypdf>=6.1.2",
    "python-dotenv>=1.1.1",
    "pyzotero>=1.7.3",
    "qdrant-client>=1.15.1",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
]
